did another refactor of the models to get rid of unreal dependencies

+++ Source/Daedalus/Models/Terrain/BiomeRegionData.cpp
@@ -5,8 +5,8 @@
 
 namespace terrain {
 	BiomeRegionData::BiomeRegionData(
-		const uint16 buffer,
-		const uint32 biomeSize,
+		const uint16_t buffer,
+		const uint32_t biomeSize,
 		const BiomeRegionOffsetVector & biomeOffset
 	) : BiomeGridSize(biomeSize),
 		BiomeOffset(biomeOffset),
@@ -23,8 +23,8 @@ namespace terrain {
 		NeighboursMerged.Set(1, 1, true);
 	}
 
-	uint64 BiomeRegionData::AddBiome(
-		const uint64 x, const uint64 y,
+	uint64_t BiomeRegionData::AddBiome(
+		const uint32_t x, const uint32_t y,
 		const BiomeCellVertex & position
 	) {
 		auto id = GetNextId();
@@ -33,22 +33,22 @@ namespace terrain {
 		return id;
 	}
 
-	std::tuple<uint64, BiomeRegionGridVector, double> BiomeRegionData::FindNearestPoint(
+	std::tuple<uint64_t, BiomeRegionGridVector, double> BiomeRegionData::FindNearestPoint(
 		utils::Vector2<> offset
 	) const {
 		double xpos = offset.X * BiomeGridSize;
 		double ypos = offset.Y * BiomeGridSize;
 
-		uint16 xstart = std::max((uint32) std::floor(xpos) - 2, 0u);
-		uint16 xend = std::min((uint32) std::ceil(xpos) + 2, BiomeGridSize - 1);
-		uint16 ystart = std::max((uint32) std::floor(ypos) - 2, 0u);
-		uint16 yend = std::min((uint32) std::ceil(ypos) + 2, BiomeGridSize - 1);
+		uint16_t xstart = std::max((uint32_t) std::floor(xpos) - 2, 0u);
+		uint16_t xend = std::min((uint32_t) std::ceil(xpos) + 2, BiomeGridSize - 1);
+		uint16_t ystart = std::max((uint32_t) std::floor(ypos) - 2, 0u);
+		uint16_t yend = std::min((uint32_t) std::ceil(ypos) + 2, BiomeGridSize - 1);
 
 		double min = 5.0;
-		uint64 vid = 0;
+		uint64_t vid = 0;
 		BiomeRegionGridVector gpos;
-		for (uint16 x = xstart; x <= xend; x++) {
-			for (uint16 y = ystart; y <= yend; y++) {
+		for (uint16_t x = xstart; x <= xend; x++) {
+			for (uint16_t y = ystart; y <= yend; y++) {
 				for (auto id : PointDistribution.Get(x, y).PointIds) {
 					double dist = (offset - Biomes.at(id)->GetLocalPosition()).Length2();
 					if (dist < min) {
@@ -64,7 +64,7 @@ namespace terrain {
 	}
 
 	void BiomeRegionData::GenerateDelaunayGraph() {
-		std::vector<std::pair<BiomeCellVertex, uint64> > vertexList;
+		std::vector<std::pair<BiomeCellVertex, uint64_t> > vertexList;
 		for (auto & pair : Biomes)
 			vertexList.push_back(std::make_pair(pair.second->GetLocalPosition(), pair.first));
 		utils::BuildDelaunay2D(DelaunayGraph, vertexList);
